cleaned up heap files

+++ include/stdint.h
@@ -12,7 +12,7 @@ typedef long long int int64_t;
 typedef unsigned char uint8_t;
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long int uint64_t;
+typedef unsigned long int uint64_t;
 
 // Minimum-width integer types
 typedef int8_t int_least8_t;
